+++ Tests/ProjectSpecTests/SpecLoadingTests.swift
@@ -756,6 +756,7 @@ class SpecLoadingTests: XCTestCase {
                     "build": [
                         "parallelizeBuild": false,
                         "buildImplicitDependencies": false,
+                        "runPostActionsOnFailure": true,
                         "targets": [
                             "Target1": "all",
                             "Target2": "testing",
@@ -813,6 +814,7 @@ class SpecLoadingTests: XCTestCase {
 
                 try expect(scheme.build.parallelizeBuild) == false
                 try expect(scheme.build.buildImplicitDependencies) == false
+                try expect(scheme.build.runPostActionsOnFailure) == true
 
                 let expectedRun = Scheme.Run(
                     config: "debug",
@@ -981,6 +983,7 @@ class SpecLoadingTests: XCTestCase {
                             "build": [
                                 "parallelizeBuild": false,
                                 "buildImplicitDependencies": false,
+                                "runPostActionsOnFailure": true,
                                 "targets": [
                                     "Target${name_1}": "all",
                                     "Target2": "testing",
@@ -1046,6 +1049,7 @@ class SpecLoadingTests: XCTestCase {
 
                 try expect(scheme.build.parallelizeBuild) == false
                 try expect(scheme.build.buildImplicitDependencies) == false
+                try expect(scheme.build.runPostActionsOnFailure) == true
 
                 try expect(scheme.run?.storeKitConfiguration) == "Configuration.storekit"