+++ Sources/XcodeGen/main.swift
@@ -3,6 +3,6 @@ import ProjectSpec
 import XcodeGenCLI
 import Version
 
-let version = Version("2.21.1")
+let version = Version("2.22.0")
 let cli = XcodeGenCLI(version: version)
 cli.execute()