+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Next Version
 
-## 2.21.1
+## 2.22.0
 
 #### Added
 - Support `runPostActionsOnFailure` for running build post scripts on failing build [#1075](https://github.com/yonaskolb/XcodeGen/pull/1075) @freddi-kit
@@ -12,7 +12,7 @@
 - Fixed scheme config variant lookups for some configs like `ProdDebug` and `Prod-Debug` that broke in 2.21.0 [#1070](https://github.com/yonaskolb/XcodeGen/pull/1070) @yonaskolb
 - Fixed Xcode alerting to project changes after regeneration [#1072](https://github.com/yonaskolb/XcodeGen/pull/1072) @yonaskolb
 
-[Commits](https://github.com/yonaskolb/XcodeGen/compare/2.21.0...2.21.1)
+[Commits](https://github.com/yonaskolb/XcodeGen/compare/2.21.0...2.22.0)
 
 ## 2.21.0